chrome/pyproject.toml
[tool.poetry]
name = "chrome"
version = "0.1.0"
description = ""
authors = ["Ayan Bandyopadhyay"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
playwright = "^1.47.0"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
requests = "^2.32.3"
websockets = "^13.1"
poetry-dotenv-plugin = "^0.2.0"

[tool.poetry.scripts]
start = "chrome:start"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
